requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
Brotli==1.1.0
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cssselect2==0.8.0
Django==4.2.21
django-celery-beat==2.8.0
django-phonenumber-field==8.1.0
django-timezone-field==7.1
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
fonttools==4.58.0
frozenlist==1.6.0
idna==3.10
kombu==5.5.3
multidict==6.4.3
openpyxl==3.1.5
pdfkit==1.0.0
phonenumbers==9.0.5
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.1
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==6.0.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
twilio==9.6.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
yarl==1.20.0
zopfli==0.2.3.post1
